Refactors LineIndicator into separate file
