+++ src/components/LineIndicator.css
@@ -1,0 +1,9 @@
+.line-indicator {
+    position: absolute;
+    top: 8%;
+    background-color: #f00;
+    width: 100%;
+    height: 1px;
+    margin-left: -8px;
+    margin-top: auto;
+}
