+++ src/components/TimeSlot.css
@@ -42,15 +42,6 @@
     margin: 0.5rem;
     max-width: 100%;
 }
-.line-indicator {
-    position: absolute;
-    top: 8%;
-    background-color: #f00;
-    width: 100%;
-    height: 1px;
-    margin-left: -8px;
-    margin-top: auto;
-}
 
 @media screen and (min-width: 50rem) {
     /* Override styles for larger screens here */
